consolidate middleware directory
