+++ backend/src/middleware/authMiddleware.js
@@ -2,23 +2,17 @@ const jwt = require("jsonwebtoken");
 const User = require("../models/User");
 
 module.exports = async (req, res, next) => {
+  const token = req.headers["authorization"]?.split(" ")[1];
+  if (!token) return res.status(401).json({ message: "No token" });
+
   try {
-    const authHeader = req.headers.authorization;
-    if (!authHeader || !authHeader.startsWith("Bearer ")) {
-      return res.status(401).json({ message: "Unauthorized: No token provided." });
-    }
-
-    const token = authHeader.split(" ")[1];
-    const decoded = jwt.verify(token, process.env.JWT_SECRET || "secretkey");
-
-    const user = await User.findById(decoded.id).select("-password");
-    if (!user) {
-      return res.status(401).json({ message: "Unauthorized: User not found." });
-    }
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(decoded._id).select("-password");
+    if (!user) return res.status(401).json({ message: "User not found" });
 
     req.user = user;
     next();
-  } catch (error) {
-    res.status(401).json({ message: "Unauthorized: Invalid token." });
+  } catch {
+    res.status(401).json({ message: "Invalid token" });
   }
 };
