+++ backend/src/routes/admin.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const adminController = require('../controllers/adminController');
-const auth = require('../middlewares/authMiddleware');
-const onlyMaster = require('../middlewares/onlyMaster');
+const auth = require('../middleware/authMiddleware');
+const onlyMaster = require('../middleware/onlyMaster');
 
 router.get('/users', auth, onlyMaster, adminController.getUsers);
