+++ backend/src/routes/inventory.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const inventoryController = require('../controllers/inventoryController');
-const auth = require('../middlewares/authMiddleware');
+const auth = require('../middleware/authMiddleware');
 
 router.get('/:characterId', auth, inventoryController.getByCharacter);
 router.put('/:characterId', auth, inventoryController.update);
