+++ backend/src/routes/race.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const raceController = require('../controllers/raceController');
-const auth = require('../middlewares/authMiddleware');
+const auth = require('../middleware/authMiddleware');
 
 // Тільки для майстра або адміна
 const onlyMaster = (req, res, next) => {
